Move other-modules into internal modules
This closes #184
