+++ fast-logger/System/Log/FastLogger/File.hs
@@ -12,7 +12,7 @@ import Data.ByteString.Char8 (unpack)
 import System.Directory (doesFileExist, doesDirectoryExist, getPermissions, writable, renameFile)
 import System.FilePath (takeDirectory, dropFileName, takeFileName, (</>))
 
-import System.Log.FastLogger.Imports
+import System.Log.FastLogger.Internal.Imports
 import System.Log.FastLogger.Types (TimeFormat, FormattedTime)
 
 -- | The spec for logging files
