+++ fast-logger/System/Log/FastLogger/Internal/FileIO.hs
@@ -1,4 +1,4 @@
-module System.Log.FastLogger.FileIO where
+module System.Log.FastLogger.Internal.FileIO where
 
 import Foreign.Ptr (Ptr)
 import GHC.IO.Device (close)
@@ -6,7 +6,7 @@ import GHC.IO.FD (openFile, stderr, stdout,  writeRawBufferPtr)
 import qualified GHC.IO.FD as POSIX (FD(..))
 import GHC.IO.IOMode (IOMode(..))
 
-import System.Log.FastLogger.Imports
+import System.Log.FastLogger.Internal.Imports
 
 type FD = POSIX.FD
